This is the Human Vs Goblin Project 6
In this game we can choose a 2D grid size ranging between 1 and 10.
You can choose a certain number of humans and goblins.
The grid shows the number, symbol and health for the humans (%)and goblins (€).
The characters have to wait for their turn to move. The human player can enter the direction for each of the humans (n/s/e/w).
The goblins will move around chasing the humans, when its their turn.
IN every turn if the humans and goblins come in contact in consecutive cells horizontally or vertically then they enter in to combat which affects the health.
The games continues till either all Humans or all Goblins die